deadzone +deadzone.py implementation and doc, untested, +util.py
Implemented the deadzone ems along with basic documentation.
Also a util.py file is introduced that contains functionality (that will
be) shared by other modules. For now, this is a function that fills a
dict with missing std keys and values and a function that extracts a
subdict from a superdict.

On branch develop
Changes to be committed:
      modified:   README.md
      modified:   hessems/deadzone.py
      new file:   hessems/util.py