dev *architecture change - removed sim part from project, separated
The original plan was to provide a small simulation framework to test
and showcase the implemented EMS. However, I realized that the
simulation framework will be larger than the EMS itself, cluttering the
whole codebase. Therefore, I changed the README accordingly to reflect
this decision and removed all dummy files which are unneccessary now.

On branch develop
Changes to be committed:
      modified:   README.md
      deleted:    hessems/experiment.py
      deleted:    hessems/mockup.py
      deleted:    hessems/reference.py
      deleted:    hessems/simulate.py
      deleted:    hessems/timeseries.py
      deleted:    hessems/visualize.py